umproved qudi
Power calibration is now saving its calibration, night pause is working, nucOPS has a variable to pause the measurement, and a function that allows to create a new nuclear spin with the associated hyperfine and files. transition tracker now also automatically loads the last created spin. If one wants to change the current spin, you need to change the name in log/current_spin.csv